requirements.txt
asgiref==3.5.2
astroid==2.5.3
backports.zoneinfo==0.2.1
colorama==0.4.4
Django==4.1
django-crispy-forms==1.14.0
django-etc==1.3.0
django-resized==0.3.11
django-taggit==3.0.0
django-tinymce==3.3.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
path==16.4.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pylint==2.7.4
pystache==0.6.0
python-decouple==3.6
pytz==2021.1
sqlparse==0.4.2
static==1.1.1
toml==0.10.2
wrapt==1.12.1
